src/backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
pandas>=2.0.0
pyarrow>=10.0.0
